+++ CHANGELOG.md
@@ -1,8 +1,18 @@
 Changelog
 =========
 
-## [0.1.0] - 2018-10-07
+## [0.2.0] - 2018-10-09
+### Added
+- Completion provider for view files
+- Reference contributor for view files
+- Inspection "Returned view not used"
+- Inspection "Simplifiable view name"
+- Inspection "View does not exist"
+
+## 0.1.0 - 2018-10-07
 ### Added
 - Completion provider for translations
 - Reference contributor for translations
 - Folding builder for translations
+
+[0.2.0]: https://github.com/martynassateika/CodeIgniter-phpstorm-plugin/compare/0.1.0...0.2.0